Update i18n-messages.md
Added a note of the requirement for the I18nFilter in order for template messages to work. I spent some time digging trying to find why my messages were appearing blank. This was because I had disabled the filter and so was not setting the currentLocale render arg.
I wonder if the default behavior (template.go#L125) should be changed to return an unknown string ("??? %s ???") instead of blank?